moves WebClient to a top level source file (not inside clients dir)
